requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
langchain>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.5.0
python-dateutil>=2.8.2
numpy>=1.24.0
langgraph>=0.0.30
langchain-core>=0.1.30 